refactor: BaseEntity 추상 클래스 생성 및 각 엔티티들에 상속 처리
